+++ HISTORY.rst
@@ -3,7 +3,11 @@
 History
 -------
 
+0.1.1 (2016-02-20)
+---------------------
+* Added the ability to extract/insert fields in DCM files.
+* Added the ability to explicitly exclude fields from extraction/insertion.
+
 0.1.0 (2016-01-29)
 ---------------------
-
 * First release on PyPI.